legg til caching på barnehagebarn og flytt paginering til egen service
