+++ src/main/kotlin/no/nav/familie/ks/sak/config/CacheConfig.kt
@@ -62,4 +62,20 @@ class CacheConfig {
                 return ConcurrentMapCache(name, concurrentMap, true)
             }
         }
+
+    @Bean("barnehagelisterCache")
+    fun barnehagelisterCache(): CacheManager =
+        object : ConcurrentMapCacheManager() {
+            override fun createConcurrentMapCache(name: String): Cache {
+                val concurrentMap =
+                    Caffeine
+                        .newBuilder()
+                        .maximumSize(1000)
+                        .expireAfterWrite(24, TimeUnit.HOURS)
+                        .recordStats()
+                        .build<Any, Any>()
+                        .asMap()
+                return ConcurrentMapCache(name, concurrentMap, true)
+            }
+        }
 }
